Feat: change logic to match non-modded characters, replace regex with split()
- Replace regex with split() function to improve performance
- Refactor variable names to improve readability and meaningfulness
- Move constants and other config variables to config.py file
- Decrease number of variables by deleting the unnecessary ones

BREAKING CHANGE: change the logic to match data

The data matching will now only match these that are within the app's folder (RichPresenceData).
It will now be necessary to add a .ini file with the desired data inside that folder.
This happens so the app can match non-modded characters, and it seems to be simpler, probably increasing performance